+++ Modules/Professions/Events/Content/AssetRequested/ProfessionAssetRequestedEvent.cs
@@ -42,8 +42,11 @@ internal sealed class ProfessionAssetRequestedEvent : AssetRequestedEvent
             $"{Manifest.UniqueID}/MaxIcon",
             new ModTextureProvider(() => "assets/sprites/interface/max.png"));
         this.Provide(
-            $"{Manifest.UniqueID}/PrestigeRibbon",
+            $"{Manifest.UniqueID}/PrestigeRibbons",
             new ModTextureProvider(() => $"assets/sprites/interface/{ProfessionsModule.Config.Prestige.Ribbon}.png"));
+        this.Provide(
+            $"{Manifest.UniqueID}/PrestigeIcons",
+            new ModTextureProvider(() => $"assets/sprites/interface/professions_{ProfessionsModule.Config.Prestige.Icon}.png"));
         this.Provide(
             $"{Manifest.UniqueID}/SkillBars",
             new ModTextureProvider(ProvideSkillBars));
@@ -435,7 +438,7 @@ internal sealed class ProfessionAssetRequestedEvent : AssetRequestedEvent
     private static void EditCursorsLooseSprites(IAssetData asset)
     {
         var editor = asset.AsImage();
-        var sourceTx = ModHelper.ModContent.Load<Texture2D>("assets/sprites/interface/professions");
+        var sourceTx = ModHelper.GameContent.Load<Texture2D>($"{Manifest.UniqueID}/PrestigeIcons");
         var sourceArea = new Rectangle(0, 0, 96, 80);
         var targetArea = new Rectangle(0, 624, 96, 80);
         editor.PatchImage(sourceTx, sourceArea, targetArea);