+++ Modules/Professions/Patchers/Prestige/Integration/NewSkillsPagePerformHoverActionPatcher.cs
@@ -75,7 +75,7 @@ internal sealed class NewSkillsPagePerformHoverActionPatcher : HarmonyPatcher
                 _ => Skill.FromValue(i),
             };
 
-            var professionsForThisSkill = Game1.player.GetProfessionsForSkill(skill, true);
+            var professionsForThisSkill = Game1.player.GetProfessionsForSkill(skill, false);
             var count = professionsForThisSkill.Length;
             if (count == 0)
             {
@@ -126,7 +126,7 @@ internal sealed class NewSkillsPagePerformHoverActionPatcher : HarmonyPatcher
             bounds.Y += 56;
 
             var professionsForThisSkill =
-                Game1.player.GetProfessionsForSkill(skill, true);
+                Game1.player.GetProfessionsForSkill(skill, false);
             var count = professionsForThisSkill.Length;
             if (count == 0)
             {