+++ Modules/Ponds/Patchers/FishingRodPullFishFromWaterPatcher.cs
@@ -73,7 +73,6 @@ internal sealed class FishingRodPullFishFromWaterPatcher : HarmonyPatcher
 
     #region handlers
 
-    // ReSharper disable once RedundantAssignment
     private static void PullAlgae(FishPond pond, ref int index, ref int quality)
     {
         quality = SObject.lowQuality;
@@ -163,7 +162,6 @@ internal sealed class FishingRodPullFishFromWaterPatcher : HarmonyPatcher
         }
     }
 
-    // ReSharper disable once RedundantAssignment
     private static void HandleLegendary(FishPond pond, ref int index, ref int quality, List<int> fishQualities, int lowestFish)
     {
         var familyCount = pond.Read<int>(DataKeys.FamilyLivingHere);