+++ Modules/Professions/Patchers/GameLocationShowQiCatInnerPatcher.cs
@@ -33,7 +33,6 @@ internal sealed class GameLocationShowQiCatInnerPatcher : HarmonyPatcher
 
         if (ProfessionsModule.EnablePrestigeLevels)
         {
-            // ReSharper disable once RedundantAssignment
             __result = Math.Min(
                 Skill.ListVanilla
                     .Where(skill => skill.CurrentLevel >= 20)
@@ -45,7 +44,6 @@ internal sealed class GameLocationShowQiCatInnerPatcher : HarmonyPatcher
 
         if (ProfessionsModule.EnableSkillReset)
         {
-            // ReSharper disable once RedundantAssignment
             __result += Math.Min(
                 Skill.ListVanilla
                     .Where(skill => Game1.player.HasAllProfessionsInSkill(skill))