+++ Modules/Professions/Patchers/UtilityPercentGameCompleteInnerPatcher.cs
@@ -23,7 +23,6 @@ internal sealed class UtilityPercentGameCompleteInnerPatcher : HarmonyPatcher
 
     /// <summary>Patch to add new perfection requirement.</summary>
     [HarmonyPrefix]
-    // ReSharper disable once UnusedParameter.Local
     private static bool UtilityPercentGameCompletePrefix(ref float __result)
     {
         if (!ProfessionsModule.EnablePrestige || !ProfessionsModule.Config.Prestige.IsPerfectionRequirement)
@@ -33,7 +32,6 @@ internal sealed class UtilityPercentGameCompleteInnerPatcher : HarmonyPatcher
 
         if (ProfessionsModule.EnablePrestigeLevels)
         {
-            // ReSharper disable once RedundantAssignment
             __result = Math.Min(
                 Skill.ListVanilla
                     .Where(skill => skill.CurrentLevel >= 20)
@@ -45,7 +43,6 @@ internal sealed class UtilityPercentGameCompleteInnerPatcher : HarmonyPatcher
 
         if (ProfessionsModule.EnableSkillReset)
         {
-            // ReSharper disable once RedundantAssignment
             __result += Math.Min(
                 Skill.ListVanilla
                     .Where(skill => Game1.player.HasAllProfessionsInSkill(skill))